Corrected spelling mistakes stide --> stride and columb --> column throughout.
Gone some way towards integrating image tracker with mosaicing code to
be able to do registration of new images to a mosaic, enabling foreground
separation.

Fixed a couple of bugs in the image tracker code: patch around corner needs
to be centred on zero and the corners need to be computed from the image,
not the mosaic.


git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl@1888 7226f50c-d813-0410-99bf-979b288ab488